Use baasaas on CI (#1441)
* Use baasaas on CI

* Remove dependency on realm in the baas client

* Regenerate models

* Cleanup containers after test runs

* Fix CI

* Try setting the baas url correctly

* Fix order of test setup

* ..

* Add some missing flutter tests

* Try to fix flutter tests

* Rework indexed test to use values from the second half of the table.

* Use api key auth

* Post-merge fixes

* Use endpoints, wire-up differentiator tag

* Guard against incomplete containers

* Fix cleanup workflow; fix env variable names

* Don't log isolate warnings when the isolate is created by dart test

* Pass correct differentiator to android tests

* Always run cleanup

* Rework cleanup command

* Use reporter

* Don't fail test run on test errors

* Fail on error

* Remove failing test

* Update lib/src/cli/atlas_apps/options.dart


* Support maps (#1406)

* Update generator to support maps

* Generator updates

* Better error message for wrong key type

* Remove the default value lists

* Add tests for non-empty default collection initializers

* Wire up some of the implementation

* Wire up most of the test infrastructure

* Add notification tests

* Fix generator test expects

* Fix tests

* Add changelog, clean up the generator a little

* Revert some unneeded changes

* Fix expectations

---------


* Merge main, regenerate cli

---------

+++ lib/src/cli/atlas_apps/deleteapps_command.dart
@@ -56,13 +56,22 @@ class DeleteAppsCommand extends Command<void> {
         abort('--project-id must be supplied when --atlas-cluster is not set');
       }
     }
+
+    if (options.baasaasApiKey == null && options.baasUrl == null) {
+      abort('--baas-url must be supplied when --baasaas-api-key is null');
+    }
+
     final differentiator = options.differentiator ?? 'local';
 
-    final client = await (options.atlasCluster == null
-        ? BaasClient.docker(options.baasUrl, differentiator)
-        : BaasClient.atlas(options.baasUrl, options.atlasCluster!, options.apiKey!, options.privateApiKey!, options.projectId!, differentiator));
+    if (options.baasaasApiKey != null) {
+      await BaasClient.retry(() => BaasClient.deleteContainer(options.baasaasApiKey!, differentiator));
+    } else {
+      final client = await (options.atlasCluster == null
+          ? BaasClient.docker(options.baasUrl!, differentiator)
+          : BaasClient.atlas(options.baasUrl!, options.atlasCluster!, options.apiKey!, options.privateApiKey!, options.projectId!, differentiator));
 
-    await client.deleteApps();
+      await client.deleteApps();
+    }
   }
 
   void abort(String error) {
